update CuGraphGATConv to use base class
