+++ torch_geometric/nn/conv/cugraph/__init__.py
@@ -1,7 +1,9 @@
 from .base import CuGraphModule
 from .sage_conv import CuGraphSAGEConv
+from .gat_conv import CuGraphGATConv
 
 __all__ = [
     'CuGraphModule',
     'CuGraphSAGEConv',
+    'CuGraphGATConv',
 ]
