+++ torch_geometric/nn/conv/cugraph/gat_conv.py
@@ -1,26 +1,25 @@
 from typing import Optional, Tuple
 
 import torch
-import torch.nn.functional as F
-from torch import Tensor
-from torch.nn import Linear
+from torch import Tensor, nn
 
 from torch_geometric.nn.conv.cugraph import CuGraphModule
+from torch_geometric.nn.inits import zeros
 
 try:
     from pylibcugraphops import make_fg_csr, make_mfg_csr
-    from pylibcugraphops.torch.autograd import agg_concat_n2n as SAGEConvAgg
+    from pylibcugraphops.torch.autograd import mha_gat_n2n as GATConvAgg
     HAS_PYLIBCUGRAPHOPS = True
 except ImportError:
     HAS_PYLIBCUGRAPHOPS = False
 
 
-class CuGraphSAGEConv(CuGraphModule):  # pragma: no cover
-    r"""The GraphSAGE operator from the `"Inductive Representation Learning on
-    Large Graphs" <https://arxiv.org/abs/1706.02216>`_ paper.
+class CuGraphGATConv(CuGraphModule):  # pragma: no cover
+    r"""The graph attentional operator from the `"Graph Attention Networks"
+    <https://arxiv.org/abs/1710.10903>`_ paper.
 
-    :class:`CuGraphSAGEConv` is an optimized version of
-    :class:`~torch_geometric.nn.conv.SAGEConv` based on the :obj:`cugraph-ops`
+    :class:`CuGraphGATConv` is an optimized version of
+    :class:`~torch_geometric.nn.conv.GATConv` based on the :obj:`cugraph-ops`
     package that fuses message passing computation for accelerated execution
     and lower memory footprint.
     """
@@ -28,42 +27,42 @@ class CuGraphSAGEConv(CuGraphModule):  # pragma: no cover
         self,
         in_channels: int,
         out_channels: int,
-        aggr: str = 'mean',
-        normalize: bool = False,
-        root_weight: bool = True,
-        project: bool = False,
+        heads: int = 1,
+        concat: bool = True,
+        negative_slope: float = 0.2,
+        add_self_loops: bool = True,
         bias: bool = True,
     ):
         if HAS_PYLIBCUGRAPHOPS is False:
             raise ModuleNotFoundError(f"'{self.__class__.__name__}' requires "
                                       f"'pylibcugraphops >= 23.02.00'")
 
-        if aggr not in ['mean', 'sum', 'min', 'max']:
-            raise ValueError(f"Aggregation function must be either 'mean', "
-                             f"'sum', 'min' or 'max' (got '{aggr}')")
-
         super().__init__()
         self.in_channels = in_channels
         self.out_channels = out_channels
-        self.aggr = aggr
-        self.normalize = normalize
-        self.root_weight = root_weight
-        self.project = project
+        self.heads = heads
+        self.concat = concat
+        self.negative_slope = negative_slope
+        self.add_self_loops = add_self_loops
 
-        if self.project:
-            self.pre_lin = Linear(in_channels, in_channels, bias=True)
+        self.lin = nn.Linear(in_channels, heads * out_channels, bias=False)
+        self.att = nn.Parameter(torch.Tensor(2 * heads * out_channels))
 
-        if self.root_weight:
-            self.lin = Linear(2 * in_channels, out_channels, bias=bias)
+        if bias and concat:
+            self.bias = nn.Parameter(torch.Tensor(heads * out_channels))
+        elif bias and not concat:
+            self.bias = nn.Parameter(torch.Tensor(out_channels))
         else:
-            self.lin = Linear(in_channels, out_channels, bias=bias)
+            self.register_parameter('bias', None)
 
         self.reset_parameters()
 
     def reset_parameters(self):
-        if self.project:
-            self.pre_lin.reset_parameters()
         self.lin.reset_parameters()
+        gain = nn.init.calculate_gain('relu')
+        nn.init.xavier_normal_(self.att.view(2, self.heads, self.out_channels),
+                               gain=gain)
+        zeros(self.bias)
 
     def forward(
         self,
@@ -75,12 +74,9 @@ class CuGraphSAGEConv(CuGraphModule):  # pragma: no cover
             raise RuntimeError(f"'{self.__class__.__name__}' requires GPU-"
                                f"based processing (got CPU tensor)")
 
-        if self.project:
-            x = self.pre_lin(x).relu()
-
         row, colptr = csc
 
-        # Create `cugraph-ops` graph:
+        # Create `cugraph-ops`` graph:
         if x.size(0) != colptr.numel() - 1:  # Operating in a bipartite graph:
             if max_num_neighbors is None:
                 max_num_neighbors = int((colptr[1:] - colptr[:-1]).max())
@@ -93,18 +89,15 @@ class CuGraphSAGEConv(CuGraphModule):  # pragma: no cover
         else:
             _graph = make_fg_csr(colptr, row)
 
-        out = SAGEConvAgg(x, _graph, self.aggr)
+        x = self.lin(x)
+        out = GATConvAgg(x, self.att, _graph, self.heads, "LeakyReLU",
+                         self.negative_slope, False, self.concat)
 
-        if self.root_weight:
-            out = self.lin(out)
-        else:
-            out = self.lin(out[:, :self.in_channels])
-
-        if self.normalize:
-            out = F.normalize(out, p=2., dim=-1)
+        if self.bias is not None:
+            out = out + self.bias
 
         return out
 
     def __repr__(self) -> str:
         return (f'{self.__class__.__name__}({self.in_channels}, '
-                f'{self.out_channels}, aggr={self.aggr})')
+                f'{self.out_channels}, heads={self.heads})')
